% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_err.R
\name{calc_err}
\alias{calc_err}
\title{TO DO}
\usage{
calc_err(traj, gps_data, baro_data, printMaxErrors = TRUE)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
