% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sensor.R
\name{check_sensor_model}
\alias{check_sensor_model}
\title{TO DO}
\usage{
check_sensor_model(name, mdl)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
