% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_traj.R
\name{check_traj}
\alias{check_traj}
\title{TO DO}
\usage{
check_traj(traj, timing)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
