% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{cnstr.e_gyr}
\alias{cnstr.e_gyr}
\title{TO DO}
\usage{
cnstr.e_gyr(X)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
