% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{cnstr.IC}
\alias{cnstr.IC}
\title{TO DO}
\usage{
cnstr.IC(traj.ref, KF.mdl)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
