% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_traj.R
\name{complete_traj}
\alias{complete_traj}
\title{TO DO}
\usage{
complete_traj(traj)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
