% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Compute Coverage}
\usage{
compute_coverage(sols, alpha = 0.95, step = 100, idx = 1:6)
}
\arguments{
\item{sols}{The set of solutions returned by the \code{navigation} function}

\item{alpha}{size of the confidence interval}

\item{step}{Step}

\item{idx}{Components of the states to be considered (default: position and orientation)}
}
\description{
Compute Empirical Coverage
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
