% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_coverage_imu_states}
\alias{compute_coverage_imu_states}
\title{Compute Coverage for the IMU states}
\usage{
compute_coverage_imu_states(sols, alpha = 0.95, step = 100, idx = 1:6)
}
\arguments{
\item{sols}{The set of solutions returned by the \code{navigation} function}

\item{alpha}{size of the confidence interval}

\item{step}{Step.}

\item{idx}{Components to be considered (idx in [1, ..., 6])}
}
\description{
works correctly only if there is only one process that is not a WN
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
