% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_mean_position_err}
\alias{compute_mean_position_err}
\title{Compute mean position error}
\usage{
compute_mean_position_err(sols, step = 1, t0 = NULL, tend = NULL)
}
\arguments{
\item{sols}{The set of solutions returned by the \code{navigation} function}

\item{step}{do it for one sample out of \code{step}}

\item{t0}{Start time for RMS calculation (default: beginning)}

\item{tend}{Start time for RMS calculation (default: end)}
}
\description{
Compute the mean position error (norm of 3D NED error)
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
