% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{compute_rms}
\alias{compute_rms}
\title{Compute Root Mean Squared error (RMS)}
\usage{
compute_rms(sols, step = 1, idx = 1:3, t0 = NULL, tend = NULL)
}
\arguments{
\item{sols}{The set of solutions returned by the \code{navigation} function}

\item{step}{do it for one sample out of \code{step}}

\item{idx}{Components of the states to be considered (default: position)}

\item{t0}{Start time for RMS calculation (default: beginning)}

\item{tend}{Start time for RMS calculation (default: end)}
}
\description{
Compute Root Mean Squared error (RMS) for each solution
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
