% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{EKF.pred}
\alias{EKF.pred}
\title{TO DO}
\usage{
EKF.pred(X_k, P_k, Fee, Wmat, imu_data, Betas, g, dt, method)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
