% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{example_3_traj_ned}
\alias{example_3_traj_ned}
\title{Example trajectory 3 in NED coordinates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 18001 rows and 7 columns.
}
\usage{
example_3_traj_ned
}
\description{
Example trajectory 3 in NED coordinates.
}
\keyword{datasets}
