% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{gather_GM_betas}
\alias{gather_GM_betas}
\title{collects GM betas from ts.model into one 6x5 matrix}
\usage{
gather_GM_betas(mdl.acc, mdl.gyr)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
