% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_snsr_data.R
\name{gen_snsr_data}
\alias{gen_snsr_data}
\title{TO DO}
\usage{
gen_snsr_data(
  snsr.name,
  traj,
  timing,
  snsr.mdl = NULL,
  g = NULL,
  clean_data = FALSE
)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
