% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_time}
\alias{get_time}
\title{Extract time for trajectory object}
\usage{
get_time(x)
}
\arguments{
\item{x}{A \code{trajectory} object.}
}
\value{
A \code{numeric} vector with the time in seconds of a trajectory object.
}
\description{
Extract time for trajectory object in seconds
}
\examples{
\dontrun{
n = 1000
dat = cbind(seq(from = 0, to = 50, length.out = n),
 46.204391 + cumsum(rnorm(n))/100, 
 6.143158 + cumsum(rnorm(n))/100,
  375 + cumsum(rnorm(n)))
traj = make_trajectory(data = dat, name = "My cool data")
get_time(traj)

}

}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
