% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{INS_GPS_EKF}
\alias{INS_GPS_EKF}
\title{TO DO}
\usage{
INS_GPS_EKF(
  IC,
  imu_data,
  gps_data = NULL,
  baro_data = NULL,
  KF.mdl = NULL,
  g,
  noProgressBar = FALSE,
  PhiQ_method = "exact",
  P_subsampling = 1
)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
