% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_mat.R
\name{KF_mat.Fee}
\alias{KF_mat.Fee}
\title{TO DO}
\usage{
KF_mat.Fee(mdl.acc, mdl.gyr)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
