% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KF_mat.R
\name{KF_mat.Rmat}
\alias{KF_mat.Rmat}
\title{TO DO}
\usage{
KF_mat.Rmat(mdl.gps = NULL, mdl.baro = NULL)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
