% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{lemniscate_traj_ned}
\alias{lemniscate_traj_ned}
\title{Lemniscate trajectory in NED coordinates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 60001 rows and 7 columns.
}
\usage{
lemniscate_traj_ned
}
\description{
Lemniscate trajectory in NED coordinates.
}
\keyword{datasets}
