% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manipulations.R
\name{model_to_ekf}
\alias{model_to_ekf}
\title{stochastic model conversion for the EFK}
\usage{
model_to_ekf(model, frequency)
}
\arguments{
\item{model}{The model to be converted.}

\item{frequency}{The frequency at which this model is supposed to be sampled in Hz.}
}
\value{
a \code{ts.model} suitable to be used in \code{INS_GPS_EKF}.
}
\description{
This function converts a discrete time \code{ts.model} to the equivalent continuous time representation, suitable to be employed in \code{INS_GPS_EKF}
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
