% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{navigation}
\alias{navigation}
\title{Runs "IMU model evaluation" or "INS-GPS-Baro integrated navigation (sensor fusion)"}
\usage{
navigation(
  traj.ref,
  timing,
  snsr.mdl,
  KF.mdl,
  g = 9.8056,
  num.runs = 1,
  results.system = "ned",
  x_o = NULL,
  noProgressBar = FALSE,
  IC = NULL,
  imu_data = NULL,
  gps_data = NULL,
  baro_data = NULL,
  input.seed = 0,
  PhiQ_method = "exact",
  P_subsampling = 1,
  compute_PhiQ_each_n = 1,
  parallel.ncores = detectCores(all.tests = FALSE, logical = TRUE),
  tmpdir = tempdir()
)
}
\arguments{
\item{traj.ref}{A \code{trajectory} object (see the documentation for \code{make_trajectory}), serving as the reference trajectory for generating sensor data and evaluating the error in navigation once performed. Only position and attitude data are required/considered, and velocity will be calculated from position.}

\item{timing}{A \code{timing} object (see the documentation for \code{make_timing}) containing timing information such as start and end of navigation.}

\item{snsr.mdl}{A \code{sensor} object (see the documentation for \code{make_sensor}) containing additive sensor error model to generate realistic sensor data.}

\item{KF.mdl}{A \code{sensor} object (see the documentation for \code{make_sensor}) containing additive sensor error model to be used within the Kalman filter for navigation.}

\item{g}{Gravitational acceleration.}

\item{num.runs}{Number of times the sensor data generation and navigation is performed (Monte-Carlo simulation).}

\item{results.system}{The coordinate system (\code{ned}/\code{ellipsoidal}) in which the results are reported (see the documentation for \code{make_trajectory}).}

\item{x_o}{Origin of the fixed \code{ned} frame.}

\item{noProgressBar}{A \code{bolean} specifying if there should not be a progress bar}

\item{IC}{Initial conditions. See the examples for the format.}

\item{imu_data}{IMU data. See the examples for the format.}

\item{gps_data}{GPS data. See the examples for the format.}

\item{baro_data}{Baro data. See the examples for the format.}

\item{input.seed}{Seed for the random number generator. Actual seed is computed as \code{input.seed * num.runs + run}}

\item{PhiQ_method}{String that specify the method to compute Phi and Q matrices, can be "exact" or the order of the Taylor expansions to use}

\item{P_subsampling}{(memory optimization) store only one sample of the P matrix each \code{P_subsampling} time instants}

\item{compute_PhiQ_each_n}{Specify the interval of IMU measurements between each computation of PhiQ.}

\item{parallel.ncores}{The number of cores to be used for parallel Monte-Carlo runs}

\item{tmpdir}{Where to store temporary navigation output. It should not be mapped on a filesystem which lives in RAM.}
}
\value{
An object of \code{navigation} class containing the reference trajectory, fused trajectory, sensor data, covariance matrix, and time.
}
\description{
This function performs of the two following main tasks, base on the provided input. If a reference trajectory (\code{traj.ref}) is provided, it generates sensor data (IMU, GPS, Baro) corrupted by additive errors according to \code{snsr.mdl}, and performs navigation using \code{KF.mdl} as the sensor error model within the Kalman filter to evaluate how this particular model performs when navigating.
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
