% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.coverage.stat}
\alias{plot.coverage.stat}
\title{Plot coverage}
\usage{
\method{plot}{coverage.stat}(..., legend = NA, title = NA)
}
\arguments{
\item{...}{coverage, e.g., computed with \code{compute_coverage}}

\item{legend}{Legend of the plot.}

\item{title}{Title of the plot.}
}
\description{
plot multiple coverages alltogether
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
