% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.navigation.stat}
\alias{plot.navigation.stat}
\title{Plot a navigation statistic}
\usage{
\method{plot}{navigation.stat}(..., legend = NA, title = NA, xlim = c(NA, NA), ylim = c(NA, NA))
}
\arguments{
\item{...}{navigation statistics, e.g., computed with \code{compute_mean_position_err}}

\item{legend}{The legend}

\item{title}{The title}

\item{xlim}{xlim}

\item{ylim}{ylim}
}
\description{
plot multiple stats alltogether
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
