% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.nees.stat}
\alias{plot.nees.stat}
\title{Plot multiple \code{nees.stat} objects}
\usage{
\method{plot}{nees.stat}(..., alpha = 0.95, legend = NA, title = NA)
}
\arguments{
\item{...}{NEES, e.g., computed with \code{compute_nees}}

\item{alpha}{for the confidence interval plot}

\item{legend}{legend of the plot.}

\item{title}{title of the plot.}
}
\description{
plot multiple \code{nees.stat} objects alltogether
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
