% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_coverage}
\alias{plot_coverage}
\title{plot coverage}
\usage{
plot_coverage(coverage, alpha = 0.95, nruns)
}
\arguments{
\item{coverage}{a list of result of a call to \code{compute_coverage}}

\item{alpha}{size of the confidence interval}

\item{nruns}{number of runs (to calculate CI)}
}
\description{
plot coverage
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
