% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_imu_err_with_cov}
\alias{plot_imu_err_with_cov}
\title{Plot IMU error with covariances}
\usage{
plot_imu_err_with_cov(sol, idx = 1, error = TRUE, step = 10)
}
\arguments{
\item{sol}{The set of solutions returned by the \code{navigation} function}

\item{idx}{Which Monte-Carlo solution to plot}

\item{error}{Whether to plot the error with respect to the reference or the estimated values}

\item{step}{Plot one time out of \code{step}}
}
\description{
this function plots the estimated IMU errors with covariance of a solution computed with the \code{navigation} function
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
