% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_nav_states_with_cov}
\alias{plot_nav_states_with_cov}
\title{Plot navigation states with covariance}
\usage{
plot_nav_states_with_cov(sol, idx = 1, cov_idx = 1, error = TRUE, step = 10)
}
\arguments{
\item{sol}{The set of solutions returned by the \code{navigation} function}

\item{idx}{Which Monte-Carlo solution to plot (can be a vector)}

\item{cov_idx}{Which Monte-Carlo solution to use for confidence intervals}

\item{error}{Wether to plot the error with respect to the refefence or the estimated values}

\item{step}{Plot one time out of \code{step}}
}
\description{
this function plots the navigation states with estimated covariance of a solution computed with the \code{navigation} function
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
