% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_nees}
\alias{plot_nees}
\title{Plot Normalized Estimation Errror Squared (NEES)}
\usage{
plot_nees(nees, nruns, dof, alpha = 0.95)
}
\arguments{
\item{nees}{a list of result of a call to \code{compute_nees}}

\item{nruns}{number of runs (to calculate CI)}

\item{dof}{number of variables (to calculate CI)}

\item{alpha}{to calculate CI}
}
\description{
Plot Normalized Estimation Errror Squared (NEES)
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
