% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_rms}
\alias{plot_rms}
\title{Plot Root Mean Squared error (RMS)}
\usage{
plot_rms(
  rms,
  idx = 1:3,
  t0 = NULL,
  tend = NULL,
  ylabels = c("N", "E", "D", "Roll", "Pitch", "Yaw")
)
}
\arguments{
\item{rms}{a list of results obtained from \code{compute_rms}}

\item{idx}{Which components of the RMS to plot, default (1:3, position)}

\item{t0}{start time (default: beginning)}

\item{tend}{stop time (defaut: end)}

\item{ylabels}{custom labels for the y axis (default: labels for position and orientation)}
}
\description{
this function plots the RMS computed with the \code{compute_rms} function
}
\author{
Davide Cucci, Lionel Voirol, Mehran Khaghani, Stéphane Guerrier
}
