% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{pred.PhiQ}
\alias{pred.PhiQ}
\title{TO DO}
\usage{
pred.PhiQ(Fmat, Gmat, Wmat, dt, method = "exact")
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
