% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{pred.Xn}
\alias{pred.Xn}
\title{TO DO}
\usage{
pred.Xn(X_k, imu_data, dt, g)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
