% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_mat.R
\name{rot.dC_b_i_dy}
\alias{rot.dC_b_i_dy}
\title{TO DO}
\usage{
rot.dC_b_i_dy(r, p, y)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
