% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_mat.R
\name{rot.dCw_inv_dp}
\alias{rot.dCw_inv_dp}
\title{TO DO}
\usage{
rot.dCw_inv_dp(r, p)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
