% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INS_GPS_EKF.R
\name{used_indices}
\alias{used_indices}
\title{Returns the meaningul row/column indices of the covariance matrix}
\usage{
used_indices(mdl.imu)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
