% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGS84_datum.R
\name{WGS84_datum}
\alias{WGS84_datum}
\title{TO DO}
\usage{
WGS84_datum()
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
