% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_trans.R
\name{X_ecef2ellips}
\alias{X_ecef2ellips}
\title{TO DO}
\usage{
X_ecef2ellips(X_ec, ignoreAtan2Flag = FALSE)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
