% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_trans.R
\name{X_ellips2ecef}
\alias{X_ellips2ecef}
\title{TO DO}
\usage{
X_ellips2ecef(x)
}
\arguments{
\item{TO}{DO}
}
\value{
TO DO
}
\description{
TO DO
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
