% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot.nees.stat}
\alias{plot.nees.stat}
\title{plot NEES}
\usage{
\method{plot}{nees.stat}(..., alpha = 0.95, legend = NA, title = NA)
}
\arguments{
\item{...}{NEES, e.g., computed with \code{compute_nees}}

\item{alpha}{for the confidence interval plot}

\item{legend}{legend of the plot.}

\item{title}{title of the plot.}
}
\description{
plot multiple NEES alltogether
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
