% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sensor.R
\name{print.sensor}
\alias{print.sensor}
\title{Print a \code{sensor} object parameters (name, frequency and error model)}
\usage{
\method{print}{sensor}(x)
}
\arguments{
\item{x}{A \code{sensor} object.}
}
\value{
Print the \code{sensor} object name and specifications in the console.
}
\description{
Print method for a \code{sensor} object
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
