% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sensor.R
\name{print.sensor}
\alias{print.sensor}
\title{print.sensor}
\usage{
\method{print}{sensor}(x)
}
\arguments{
\item{x}{A \code{sensor} object.}
}
\value{
Print the sensor name and specifications in the console.
}
\description{
Print method for a sensor object
}
\author{
Stephane Guerrier, Mehran Khaghani, and Lionel Voirol
}
