% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_trans.R
\name{X_ecef2ellips}
\alias{X_ecef2ellips}
\title{Convert ECEF to ellipsoidal coordinates}
\usage{
X_ecef2ellips(X_ec, ignoreAtan2Flag = FALSE)
}
\arguments{
\item{X_ec}{a 3 x n matrix containing the ECEF coordinatesto be converted}

\item{ignoreAtan2Flag}{A \code{boolean} value indicating if Atan2Flag should be ignored.}
}
\value{
a matrix containing the WGS84 latitude, longitude, altitude converted coordinates
}
\description{
Convert ECEF to ellipsoidal (latitude, longitude, altitude, in WGS84) coordinates
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
