% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_trans.R
\name{X_ecef2ned}
\alias{X_ecef2ned}
\title{Transform ECEF to NED coordinates}
\usage{
X_ecef2ned(x, x_o)
}
\arguments{
\item{x}{a 3 x n matrix with ECEF coordinates to be converted}

\item{x_o}{the latitude, longitude, altitude of the desired local level origin}
}
\value{
\code{x} converted to the local level
}
\description{
Transform ECEF to local-level NED coordinates centered at \code{x_0}
}
\author{
Stephane Guerrier, Mehran Khaghani, Lionel Voirol and Davide A. Cucci
}
